More SAST/UT Fixes

+++ sim-core/src/main/java/com/sim/chatserver/config/MapReduceConfig.java
@@ -323,7 +323,7 @@ public final class MapReduceConfig {
         if (v != null && !v.isBlank()) {
             Integer parsed = parseInt(v);
             if (parsed != null) {
-                return parsed;
+                return parsed.intValue();
             }
         }
         return defaultValue;
